fix: convert remaining identity-map MMIO/DMA accesses to high-half VMA
After identity map removal (PML4[0] = 0), all physical memory
accesses must go through the high-half VMA (PHYS_TO_VIRT / vmm_map_phys).

Drivers fixed:
- usb_ehci: convert BAR0 to PHYS_TO_VIRT for EHCI MMIO registers
- usb_msc: alloc_dma returns PHYS_TO_VIRT ptr; fix all pmm_free_frame
  and DMA buffer address derivation (use VIRT_TO_PHYS for DMA descriptors)
- ahci: map HBA BAR5 via vmm_map_phys; fix all memset/memcpy to use
  PHYS_TO_VIRT for DMA memory (command list, FIS, command table, data)
- vga: map framebuffer from PCI BAR0 via vmm_map_phys in bochs_vbe path;
  map stdvga MMIO BAR2 for endianness setup
- smp: cast ap_trampoline_start/end to uintptr_t for cppcheck

+++ src/kernel/smp.c
@@ -263,7 +263,7 @@ int smp_boot_aps(void) {
     ioapic_init();
 
     /* Copy trampoline to low memory (physical page 0x7000) */
-    size_t tramp_size = (size_t)(ap_trampoline_end - ap_trampoline_start);
+    size_t tramp_size = (size_t)((uintptr_t)ap_trampoline_end - (uintptr_t)ap_trampoline_start);
     if (tramp_size > 0x1000) {
         kprintf("[!!] SMP: trampoline too large (%u bytes)\n", (uint64_t)tramp_size);
         return 0;
